:bug: Fix(actions): bug with not awaiting a promise.all
