+++ .changeset/quiet-wasps-fail.md
@@ -1,0 +1,5 @@
+---
+"@tevm/actions": patch
+---
+
+Fixed bug in call actions not properly awaiting in forked mode when fetching a past block tag"
